DefaultVerticalModeFormInteractorSetAsDefaultTest
